Add files via upload
The script is used to check browser compliance for Edge, Chrome and Firefox. First it will query the system for the browser version against the latest version. If the installed version is not the latest version, it will update. Please note, for this to go through without a popup, it will need to be run as an administrator.